Deleting UserStorage deletes custom Dictionary
Needs testing on Ratings, and User deletion -> UserStorage deletion also
needs testing

+++ capstoneproject/tests/test_models/test_models/test_user_storage.py
@@ -7,13 +7,35 @@ from capstoneproject.models.models.word_feature import WordFeature
 
 
 class UserStorageTestClass(TestCase):
+    @classmethod
+    def setUpClass(cls):
+        cls.category1 = Category.categories.create(
+            default=True,
+            name='category1',
+            weight=1)
+        cls.word1 = Word.words.create(
+            default=True,
+            name='word1')
+        cls.word_feature1 = WordFeature.word_features.create(
+            default=True,
+            category=cls.category1,
+            strength=True,
+            weight=1)
+
+    @classmethod
+    def tearDownClass(cls):
+        Word.words.all().delete()
+        Category.categories.all().delete()
+        WordFeature.word_features.all().delete()
+
     def setUp(self):
         self.user = User.objects.create(username='user1', password='123456')
         self.user.save()
         self.user_storage = UserStorage.user_storage.get(user=self.user.id)
 
     def tearDown(self):
-        self.user_storage.delete()
+        if self.user_storage.id is not None:
+            self.user_storage.delete()
         self.user.delete()
 
     def test_autocreate(self):
@@ -40,3 +62,50 @@ class UserStorageTestClass(TestCase):
             user_features,
             default_features,
             msg="User's WordFeatures do not match default WordFeatures")
+
+    def test_delete_user_storage(self):
+        self.assertIsNotNone(
+            self.user_storage,
+            msg="UserStorage with user's pk was not created")
+        self.user_storage.delete()
+        self.assertFalse(
+            UserStorage.user_storage.filter(user=self.user.id).exists(),
+            msg="The deleted UserStorage still exists")
+
+    def test_delete_relatives(self):
+        user_word = Word.words.create(name='user_word', default=False)
+        self.user_storage.words.add(user_word)
+        user_category = Category.categories.create(
+            name='user_category', weight=1, default=False)
+        self.user_storage.categories.add(user_category)
+        user_feature = WordFeature.word_features.create(
+                category=user_category, weight=1, strength=True, default=False)
+        self.user_storage.word_features.add(user_feature)
+        user_word.word_features.add(user_feature)
+        self.assertTrue(
+            Word.words.filter(
+                id=user_word.id).exists(), msg='Word user_word should exist')
+        self.assertTrue(
+            Category.categories.filter(
+                id=user_category.id).exists(),
+            msg='Category user_category should exist')
+        self.assertTrue(
+            WordFeature.word_features.filter(
+                id=user_feature.id).exists(),
+            msg='WordFeature user_feature should exist')
+        self.user_storage.delete_relatives()
+        self.assertFalse(
+            Word.words.filter(
+                id=user_word.id).exists(),
+            msg='Word user_word was not deleted')
+        self.assertFalse(
+            Category.categories.filter(
+                id=user_category.id).exists(),
+            msg='Category user_category should not exist')
+        self.assertFalse(
+            WordFeature.word_features.filter(
+                id=user_feature.id).exists(),
+            msg='WordFeature user_feature should not exist')
+
+    def test_delete_user(self):
+        pass
